+++ doemaarwatt/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.10] - 2026-07-29
+
+### Fixed
+
+- Added exception handling for write_registers_parallel() in modbus library, in order to prevent fallback
+- Incremental log downloading and viewing, instead of continuous full downloads
+- Gzip transfer encoding of all log-related endpoints
+
+### Added
+
+- Log files are compressed into zip-file when pressing the download button
+
 ## [1.1.9] - 2026-07-29
 
 ### Added